LPS-53858 - Extract common logic from DDL and Journal to stage DDM independently.
