+++ portal-impl/src/com/liferay/portlet/dynamicdatalists/lar/DDLPortletDataHandler.java
@@ -17,13 +17,10 @@ package com.liferay.portlet.dynamicdatalists.lar;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
 import com.liferay.portal.kernel.dao.orm.DynamicQueryFactoryUtil;
-import com.liferay.portal.kernel.dao.orm.ExportActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.ProjectionFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.Property;
 import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.RestrictionsFactoryUtil;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.lar.BasePortletDataHandler;
 import com.liferay.portal.kernel.lar.PortletDataContext;
 import com.liferay.portal.kernel.lar.PortletDataHandlerBoolean;
@@ -34,7 +31,6 @@ import com.liferay.portal.kernel.lar.StagedModelType;
 import com.liferay.portal.kernel.lar.xstream.XStreamAliasRegistryUtil;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.xml.Element;
-import com.liferay.portal.util.PortalUtil;
 import com.liferay.portlet.dynamicdatalists.model.DDLRecord;
 import com.liferay.portlet.dynamicdatalists.model.DDLRecordSet;
 import com.liferay.portlet.dynamicdatalists.model.DDLRecordVersion;
@@ -43,11 +39,8 @@ import com.liferay.portlet.dynamicdatalists.model.impl.DDLRecordSetImpl;
 import com.liferay.portlet.dynamicdatalists.service.DDLRecordLocalServiceUtil;
 import com.liferay.portlet.dynamicdatalists.service.DDLRecordSetLocalServiceUtil;
 import com.liferay.portlet.dynamicdatalists.service.permission.DDLPermission;
-import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
+import com.liferay.portlet.dynamicdatamapping.lar.DDMPortletDataHandler;
 import com.liferay.portlet.dynamicdatamapping.model.DDMTemplate;
-import com.liferay.portlet.dynamicdatamapping.model.impl.DDMStructureImpl;
-import com.liferay.portlet.dynamicdatamapping.model.impl.DDMTemplateImpl;
-import com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalServiceUtil;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -72,17 +65,11 @@ public class DDLPortletDataHandler extends BasePortletDataHandler {
 				DDLRecordSet.class.getName()),
 			new PortletDataHandlerBoolean(
 				NAMESPACE, "records", true, false, null,
-				DDLRecord.class.getName()),
-			new PortletDataHandlerBoolean(
-				NAMESPACE, "data-definitions", true, false, null,
-				DDMStructure.class.getName(), DDLRecordSet.class.getName()));
+				DDLRecord.class.getName()));
 
 		XStreamAliasRegistryUtil.register(DDLRecordImpl.class, "DDLRecord");
 		XStreamAliasRegistryUtil.register(
 			DDLRecordSetImpl.class, "DDLRecordSet");
-		XStreamAliasRegistryUtil.register(
-			DDMStructureImpl.class, "DDMStructure");
-		XStreamAliasRegistryUtil.register(DDMTemplateImpl.class, "DDMTemplate");
 	}
 
 	@Override
@@ -113,23 +100,6 @@ public class DDLPortletDataHandler extends BasePortletDataHandler {
 
 		Element rootElement = addExportDataRootElement(portletDataContext);
 
-		if (portletDataContext.getBooleanParameter(
-				NAMESPACE, "data-definitions")) {
-
-			List<DDMTemplate> ddmTemplates = new ArrayList<>();
-
-			ActionableDynamicQuery ddmStructureActionableDynamicQuery =
-				getDDMStructureActionableDynamicQuery(
-					portletDataContext, ddmTemplates);
-
-			ddmStructureActionableDynamicQuery.performActions();
-
-			for (DDMTemplate ddmTemplate : ddmTemplates) {
-				StagedModelDataHandlerUtil.exportStagedModel(
-					portletDataContext, ddmTemplate);
-			}
-		}
-
 		if (portletDataContext.getBooleanParameter(NAMESPACE, "record-sets")) {
 			ActionableDynamicQuery recordSetActionableDynamicQuery =
 				DDLRecordSetLocalServiceUtil.getExportActionableDynamicQuery(
@@ -156,32 +126,6 @@ public class DDLPortletDataHandler extends BasePortletDataHandler {
 
 		portletDataContext.importPortletPermissions(
 			DDLPermission.RESOURCE_NAME);
-
-		if (portletDataContext.getBooleanParameter(
-				NAMESPACE, "data-definitions")) {
-
-			Element ddmStructuresElement =
-				portletDataContext.getImportDataGroupElement(
-					DDMStructure.class);
-
-			List<Element> ddmStructureElements =
-				ddmStructuresElement.elements();
-
-			for (Element ddmStructureElement : ddmStructureElements) {
-				StagedModelDataHandlerUtil.importStagedModel(
-					portletDataContext, ddmStructureElement);
-			}
-
-			Element ddmTemplatesElement =
-				portletDataContext.getImportDataGroupElement(DDMTemplate.class);
-
-			List<Element> ddmTemplateElements = ddmTemplatesElement.elements();
-
-			for (Element ddmTemplateElement : ddmTemplateElements) {
-				StagedModelDataHandlerUtil.importStagedModel(
-					portletDataContext, ddmTemplateElement);
-			}
-		}
 
 		if (portletDataContext.getBooleanParameter(NAMESPACE, "record-sets")) {
 			Element recordSetsElement =
@@ -220,7 +164,7 @@ public class DDLPortletDataHandler extends BasePortletDataHandler {
 		List<DDMTemplate> ddmTemplates = new ArrayList<>();
 
 		ActionableDynamicQuery ddmStructureActionableDynamicQuery =
-			getDDMStructureActionableDynamicQuery(
+			DDMPortletDataHandler.getDDMStructureActionableDynamicQuery(
 				portletDataContext, ddmTemplates);
 
 		ddmStructureActionableDynamicQuery.performCount();
@@ -235,61 +179,6 @@ public class DDLPortletDataHandler extends BasePortletDataHandler {
 			getRecordActionableDynamicQuery(portletDataContext);
 
 		recordActionableDynamicQuery.performCount();
-	}
-
-	protected ActionableDynamicQuery getDDMStructureActionableDynamicQuery(
-		final PortletDataContext portletDataContext,
-		final List<DDMTemplate> ddmTemplates) {
-
-		ExportActionableDynamicQuery exportActionableDynamicQuery =
-			DDMStructureLocalServiceUtil.getExportActionableDynamicQuery(
-				portletDataContext);
-
-		final ActionableDynamicQuery.AddCriteriaMethod addCriteriaMethod =
-			exportActionableDynamicQuery.getAddCriteriaMethod();
-
-		exportActionableDynamicQuery.setAddCriteriaMethod(
-			new ActionableDynamicQuery.AddCriteriaMethod() {
-
-				@Override
-				public void addCriteria(DynamicQuery dynamicQuery) {
-					addCriteriaMethod.addCriteria(dynamicQuery);
-
-					Property classNameIdProperty = PropertyFactoryUtil.forName(
-						"classNameId");
-
-					long classNameId = PortalUtil.getClassNameId(
-						DDLRecordSet.class);
-
-					dynamicQuery.add(classNameIdProperty.eq(classNameId));
-				}
-
-			});
-		exportActionableDynamicQuery.setPerformActionMethod(
-			new ActionableDynamicQuery.PerformActionMethod() {
-
-				@Override
-				public void performAction(Object object)
-					throws PortalException {
-
-					DDMStructure ddmStructure = (DDMStructure)object;
-
-					StagedModelDataHandlerUtil.exportStagedModel(
-						portletDataContext, ddmStructure);
-
-					try {
-						ddmTemplates.addAll(ddmStructure.getTemplates());
-					}
-					catch (SystemException se) {
-					}
-				}
-
-			});
-		exportActionableDynamicQuery.setStagedModelType(
-			new StagedModelType(
-				DDMStructure.class.getName(), DDLRecordSet.class.getName()));
-
-		return exportActionableDynamicQuery;
 	}
 
 	protected ActionableDynamicQuery getRecordActionableDynamicQuery(
